Fix/min liquidity (#72)
* adds min liq burn to avoid lp token inflation

* updates most test for min liq lock

* lint

* lint

+++ src/libraries/MathLib.sol
@@ -79,13 +79,13 @@ library MathLib {
     }
 
     // babylonian method (https://en.wikipedia.org/wiki/Methods_of_computing_square_roots#Babylonian_method)
-    function sqrt(uint x) public pure returns (uint y) {
-      uint z = (x + 1) / 2;
-      y = x;
-      while (z < y) {
-        y = z;
-        z = (x / z + z) / 2;
-      }
+    function sqrt(uint256 x) public pure returns (uint256 y) {
+        uint256 z = (x + 1) / 2;
+        y = x;
+        while (z < y) {
+            y = z;
+            z = (x / z + z) / 2;
+        }
     }
 
     /**
